requirements.txt
aiohttp
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.10.4
anyio==3.6.2
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
CacheControl==0.14.0
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click
coverage==7.2.7
cryptography==40.0.2
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.2.0
fastapi==0.111.0
fastapi-limiter==0.1.6
firebase-admin==6.5.0
frozenlist==1.4.0
fastapi-mail==1.4.1
google-api-core
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud==0.34.0
google-cloud-core==2.3.3
google-cloud-firestore==2.15.0
google-cloud-storage
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
greenlet==2.0.2
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httpcore==0.17.2
httplib2==0.22.0
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
msgpack==1.0.8
multidict==6.0.4
packaging==23.1
passlib==1.7.4
pluggy==1.0.0
proto-plus==1.23.0
protobuf==4.23.4
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.8.0
pydub==0.25.1
PyJWT==2.8.0
pyparsing==3.1.1
pytest==7.3.1
python-dotenv==1.0.0
python-jose==3.3.0
redis==4.6.0
requests-toolbelt==1.0.0
requests
runpod==1.6.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.12
tomli==2.0.1
twilio==8.10.2
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.21.1
Werkzeug
yarl==1.9.3
